add check that lumi file link exists or not
git-svn-id: svn+ssh://svn.cern.ch/reps/dirac/ILCDIRAC/trunk/ILCDIRAC@40049 4525493e-7705-40b1-a816-d608a930855b

+++ Workflow/Modules/PythiaAnalysis.py
@@ -68,10 +68,11 @@ class PythiaAnalysis(ModuleBase):
       return S_ERROR("Lumi files not found")
     
     originpath = path+"/%s.ep"%depdir
-    try:
-      os.symlink(originpath,"/tmp/%s.ep"%depdir)
-    except Exception,x:
-      return S_ERROR("Cannot sym link lumi file: %s %s"%(Exception,x))
+    if not os.path.exists("/tmp/%s.ep"%depdir):
+      try:
+        os.symlink(originpath,"/tmp/%s.ep"%depdir)
+      except Exception,x:
+        return S_ERROR("Cannot sym link lumi file: %s %s"%(Exception,x))
     #try :
     #  shutil.copy(originpath,"/tmp/")
     #except:
